springboot整合redis,存在登录不了的问题

+++ server/src/main/java/com/alimama/server/mapper/EmployerRoleMapper.java
@@ -6,13 +6,14 @@ import com.alimama.api.model.EmployerRole;
 import com.alimama.api.model.EmployerRoleExample;
 import org.apache.ibatis.annotations.Mapper;
 import org.apache.ibatis.annotations.Param;
+
 @Mapper
 public interface EmployerRoleMapper {
     int countByExample(EmployerRoleExample example);
 
     int deleteByExample(EmployerRoleExample example);
 
-    int deleteByPrimaryKey(Integer id);
+    int deleteByPrimaryKey(Long id);
 
     int insert(EmployerRole record);
 
@@ -20,7 +21,7 @@ public interface EmployerRoleMapper {
 
     List<EmployerRole> selectByExample(EmployerRoleExample example);
 
-    EmployerRole selectByPrimaryKey(Integer id);
+    EmployerRole selectByPrimaryKey(Long id);
 
     int updateByExampleSelective(@Param("record") EmployerRole record, @Param("example") EmployerRoleExample example);
